edge-computing/requirements.txt
# Edge Computing Requirements

numpy>=1.21.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
PyYAML>=6.0
asyncio
logging